Replaces amount related fields type to int64

+++ forex/forex.go
@@ -14,9 +14,9 @@ const (
 type (
 	QuoteRequest struct {
 		SourceCurrency      common.Currency `json:"source_currency,omitempty"`
-		SourceAmount        int             `json:"source_amount,omitempty"`
+		SourceAmount        int64           `json:"source_amount,omitempty"`
 		DestinationCurrency common.Currency `json:"destination_currency,omitempty"`
-		DestinationAmount   int             `json:"destination_amount,omitempty"`
+		DestinationAmount   int64           `json:"destination_amount,omitempty"`
 		ProcessingChannelId string          `json:"processing_channel_id,omitempty"`
 	}
 
@@ -24,9 +24,9 @@ type (
 		HttpMetadata        common.HttpMetadata `json:"http_metadata,omitempty"`
 		Id                  string              `json:"id,omitempty"`
 		SourceCurrency      common.Currency     `json:"source_currency,omitempty"`
-		SourceAmount        int                 `json:"source_amount,omitempty"`
+		SourceAmount        int64               `json:"source_amount,omitempty"`
 		DestinationCurrency common.Currency     `json:"destination_currency,omitempty"`
-		DestinationAmount   int                 `json:"destination_amount,omitempty"`
+		DestinationAmount   int64               `json:"destination_amount,omitempty"`
 		Rate                float64             `json:"rate,omitempty"`
 		ExpiresOn           time.Time           `json:"expires_on,omitempty"`
 		IsSingleUse         bool                `json:"is_single_use,omitempty"`
